pyproject.toml
[tool.poetry]
name = "telegram-bookmark-bot"
version = "0.1.0"
description = "Personal Telegram bookmark bot"
readme = "README.md"
authors = ["Matthew Shelley"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^22.7"
python-dotenv = "^1.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
